airflow-dbt/mlops/requirements.txt
pandas>=1.3
scikit-learn>=1.0
joblib>=1.1
google-cloud-bigquery>=3.10.0
google-auth>=2.0